split markdown link processing into util module. test WIP
