+++ frontend/src/components/elements/Alert/Alert.tsx
@@ -21,7 +21,7 @@ import classNames from "classnames"
 import ReactMarkdown from "react-markdown"
 import { Map as ImmutableMap } from "immutable"
 import { Alert as AlertProto } from "autogen/proto"
-// TODO move Markdown functions into util module.
+import { linkWithTargetBlank, linkReferenceHasParens } from "lib/markdown_util"
 
 import "./Alert.scss"
 
@@ -30,40 +30,6 @@ var ALERT_CSS_CLASS: ImmutableMap = {
   [AlertProto.Format.WARNING]: "alert-warning",
   [AlertProto.Format.INFO]: "alert-warning",
   [AlertProto.Format.SUCCESS]: "alert-success",
-}
-
-interface LinkProps {
-  href: string
-  children: ReactElement
-}
-
-interface LinkReferenceProps {
-  href: string
-  children: [ReactElement]
-}
-
-// Using target="_blank" without rel="noopener noreferrer" is a security risk:
-// see https://mathiasbynens.github.io/rel-noopener
-const linkWithTargetBlank = (props: LinkProps): ReactElement => (
-  <a href={props.href} target="_blank" rel="noopener noreferrer">
-    {props.children}
-  </a>
-)
-
-// Handle rendering a link through a reference, ex [text](href)
-// Don't convert to a link if only `[text]` and missing `(href)`
-const linkReferenceHasParens = (props: LinkReferenceProps): any => {
-  const { href, children } = props
-
-  if (!href) {
-    return children.length ? `[${children[0].props.children}]` : ""
-  }
-
-  return (
-    <a href={href} target="_blank" rel="noopener noreferrer">
-      {children}
-    </a>
-  )
 }
 
 interface Props {
