+++ frontend/src/components/elements/Markdown/Markdown.tsx
@@ -19,49 +19,13 @@ import React, { ReactNode, ReactElement } from "react"
 
 import ReactMarkdown from "react-markdown"
 import { Map as ImmutableMap } from "immutable"
+import { linkWithTargetBlank, linkReferenceHasParens } from "lib/markdown_util"
 
-// Ignoring typeScript for this module as it has no ts support
-// eslint-disable-next-line @typescript-eslint/ban-ts-ignore
-// @ts-ignore
 import htmlParser from "react-markdown/plugins/html-parser"
 
 import CodeBlock from "../CodeBlock"
 
 import "./Markdown.scss"
-
-interface LinkProps {
-  href: string
-  children: ReactElement
-}
-
-interface LinkReferenceProps {
-  href: string
-  children: [ReactElement]
-}
-
-// Using target="_blank" without rel="noopener noreferrer" is a security risk:
-// see https://mathiasbynens.github.io/rel-noopener
-const linkWithTargetBlank = (props: LinkProps): ReactElement => (
-  <a href={props.href} target="_blank" rel="noopener noreferrer">
-    {props.children}
-  </a>
-)
-
-// Handle rendering a link through a reference, ex [text](href)
-// Don't convert to a link if only `[text]` and missing `(href)`
-const linkReferenceHasParens = (props: LinkReferenceProps): any => {
-  const { href, children } = props
-
-  if (!href) {
-    return children.length ? `[${children[0].props.children}]` : ""
-  }
-
-  return (
-    <a href={href} target="_blank" rel="noopener noreferrer">
-      {children}
-    </a>
-  )
-}
 
 interface Props {
   width: number
