+++ frontend/src/lib/markdown_util.test.tsx
@@ -1,0 +1,29 @@
+/**
+ * @license
+ * Copyright 2018-2019 Streamlit Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import React, { ReactNode, ReactElement } from "react"
+
+import classNames from "classnames"
+import ReactMarkdown from "react-markdown"
+import { Map as ImmutableMap } from "immutable"
+import { linkWithTargetBlank, linkReferenceHasParens } from "markdown_util"
+
+
+describe("linkReference", () => {
+    // linkWithTargetBlank,  linkReferenceHasParens
+    expect True
+}
